Fix dependency cycle erros
