+++ src/styles.css
@@ -21,12 +21,13 @@ body {
   font-family: 'Kanit', sans-serif;
 }
 
-#temperature #city {
-  font-size: 55px;
+#temperature {
+  font-size: 58px;
+  font-weight: 700;
 }
 
-#temperature {
-  font-weight: 700;
+#city {
+  font-size: 55px;
 }
 
 #description {
@@ -98,7 +99,7 @@ body {
 
 .slider::after {
   position: absolute;
-  content: "C º";
+  content: "F º";
   height: 24px;
   width: 24px;
   right: 10px;
@@ -117,7 +118,7 @@ input:checked + .slider::before {
 }
 
 input:checked + .slider::after {
-  content: "F º";
+  content: "C º";
   left: 20px;
 }
